abstract the logic
